Add profile page with color picker and setting for round score
