+++ package.json
@@ -22,6 +22,7 @@
     "next": "14.2.10",
     "next-auth": "^5.0.0-beta.25",
     "react": "^18",
+    "react-colorful": "^5.6.1",
     "react-dom": "^18",
     "uuid": "^10.0.0"
   },
