+++ src/app/api/updateProfile/route.ts
@@ -1,16 +1,25 @@
 import { NextRequest, NextResponse } from 'next/server';
 import { getSession } from 'next-auth/react';
 import Connection from '@/lib/connection';
+import { auth } from "@/auth";
 
 export async function POST(req: NextRequest) {
-  const session = await getSession({ req });
+  const session = await auth();
 
   if (!session || !session.user) {
     return NextResponse.json({ error: 'Unauthorized' }, { status: 401 });
   }
 
+  console.log(session);
+
   const { color_red, color_green, color_blue, show_previous_round_score } = await req.json();
-  const playerId = session.user.playerId;
+  const playerId = (session.user as any).playerId;
+
+  if (!playerId) {
+    throw new Error("Incorrect playerId");
+  }
+
+  console.log("Updating profile: " + playerId);
 
   const connection = await Connection.getInstance().getConnection();
   try {
